+++ src/app/services/get-company/get-company.interface.ts
