+++ src/infra/data/mappers/company-data-mapper.ts
@@ -1,7 +1,7 @@
-import { Company } from "../domain/company";
-import { CompanyPersistence } from "../infra/database/company-persistence.interface";
+import { Company } from "@domain/company/company";
+import { CompanyPersistence } from "@infra/data/persistence-model/company-persistence";
 
-export class CompanyMapper {
+export class CompanyDataMapper {
   public static toPersistence(data: Company): CompanyPersistence {
     return {
       id: data._id,
