+++ src/infra/data/repositories/prisma/prisma-company-repository.ts
@@ -1,10 +1,8 @@
 import { PrismaDatabase } from "@infra/data/databases/prisma/config/prisma.database";
-import {
-  CompanyRepository,
-  CompanyUniqueArgs,
-} from "@shared/modules/companies/app/ports/company-repository";
-import { Company } from "@shared/modules/companies/domain/company";
-import { CompanyMapper } from "@shared/modules/companies/mappers/company-mapper";
+
+import { Company } from "@domain/company/company";
+import { CompanyRepository, CompanyUniqueArgs } from "@app/ports/repositories/company-repository";
+import { CompanyDataMapper } from "@infra/data/mappers/company-data-mapper";
 
 export class PrismaCompanyRepository implements CompanyRepository {
   private readonly prisma: PrismaDatabase;
@@ -24,11 +22,11 @@ export class PrismaCompanyRepository implements CompanyRepository {
 
     if (!rawCompany) return null;
 
-    return CompanyMapper.toDomain(rawCompany);
+    return CompanyDataMapper.toDomain(rawCompany);
   }
 
   public async save(data: Company): Promise<void> {
-    const rawCompany = CompanyMapper.toPersistence(data);
+    const rawCompany = CompanyDataMapper.toPersistence(data);
 
     await this.prisma.company.create({
       data: rawCompany,
