+++ src/infra/data/repositories/prisma/prisma-employee-repository.ts
@@ -4,7 +4,7 @@ import {
   EmployeeRepository,
 } from "@app/ports/repositories/employee-repository";
 import { Employee } from "@domain/employee/employee";
-import { EmployeeMapper } from "@shared/modules/employees/mapper/employee-mapper";
+import { EmployeeDataMapper } from "@infra/data/mappers/employee-data-mapper";
 
 export class PrismaEmployeeRepository implements EmployeeRepository {
   private readonly prisma: PrismaDatabase;
@@ -14,7 +14,7 @@ export class PrismaEmployeeRepository implements EmployeeRepository {
   }
 
   public async save(data: Employee): Promise<void> {
-    const rawEmployee = EmployeeMapper.toPersistence(data);
+    const rawEmployee = EmployeeDataMapper.toPersistence(data);
 
     await this.prisma.employee.create({
       data: rawEmployee,
@@ -32,6 +32,6 @@ export class PrismaEmployeeRepository implements EmployeeRepository {
 
     if (!rawEmployee) return null;
 
-    return EmployeeMapper.toDomain(rawEmployee);
+    return EmployeeDataMapper.toDomain(rawEmployee);
   }
 }
