+++ src/infra/http/middlewares/api-key-validator.ts
